Move status grouping to AgentConnectorsSection, revert AddConnectorDialog
The status grouping belongs on the main agent page connectors grid
(AgentConnectorsSection), not the AddConnectorDialog which isn't used.
Reverted AddConnectorDialog to its original flat grid layout.
